+++ test/Alias.spec.ts
@@ -1,8 +1,6 @@
-import { MilvusClient } from '../milvus';
+import { MilvusClient, ERROR_REASONS, ErrorCode } from '../milvus';
 import { IP } from '../const';
 import { genCollectionParams, GENERATE_NAME } from '../utils/test';
-import { ERROR_REASONS } from '../milvus/const/ErrorReason';
-import { ErrorCode } from '../milvus/types/Response';
 
 let milvusClient = new MilvusClient(IP);
 const COLLECTION_NAME = GENERATE_NAME();
