+++ test/Collection.spec.ts
@@ -1,9 +1,11 @@
-import { MilvusClient } from '../milvus';
+import {
+  MilvusClient,
+  DataType,
+  ErrorCode,
+  ShowCollectionsType,
+  ERROR_REASONS,
+} from '../milvus';
 import { IP } from '../const';
-import { DataType } from '../milvus/const/Milvus';
-import { ErrorCode } from '../milvus/types/Response';
-import { ShowCollectionsType } from '../milvus/types/Collection';
-import { ERROR_REASONS } from '../milvus/const/ErrorReason';
 import {
   genCollectionParams,
   VECTOR_FIELD_NAME,
