+++ test/Import.spec.ts
@@ -1,8 +1,6 @@
-import { MilvusClient } from '../milvus';
 import * as path from 'path';
+import { MilvusClient, DataType, ErrorCode } from '../milvus';
 import { IP } from '../const';
-import { DataType } from '../milvus/const/Milvus';
-import { ErrorCode } from '../milvus/types/Response';
 import {
   genCollectionParams,
   VECTOR_FIELD_NAME,
